test(patina): split no options api coverage
